Added SDK storage, cached token details
The SDK will now store the trader's order details in it's storage. For
now, the only supported option is the browser's local storage.

Token details are now cached, which means they are only retrieved once
per session. This should reduce the number of network calls
significantly when regularly polling for trader balances.